requirements.txt
openai~=1.91.0
google-genai~=1.21.1
